requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
openai==1.35.13
supabase==2.5.3
stripe==10.5.0
python-dotenv==1.0.1
requests>=2.31.0
gunicorn==22.0.0
